refactor(app): :recycle: wrap client component
wrap the client component with a server component adding metadata title with i18n support
