+++ src/app/[locale]/components/organisms/login/Login.module.css
@@ -77,17 +77,3 @@
         }
     }
 }
-
-/* ¿Qué hace este fragmento? No hay ningún elemento
-con este className en el componente */
-
-/* .google-login-button {
-    display: flex;
-    align-items: center;
-    padding: .5rem 1rem;
-    border: none;
-    border-radius: 9px;
-    font-size: 16px;
-    font-weight: 700;
-    gap: .5rem;
-} */
