+++ src/app/[locale]/login/page.tsx
@@ -1,61 +1,23 @@
-'use client'
+import { getTranslator } from 'next-intl/server'
 
-import Link from 'next/link'
-import Image from 'next/image'
+import Login from '@/components/organisms/login/Login'
+import { MetadataProps } from '@/app/layout.types'
 
-import { signIn } from 'next-auth/react'
-import { useSearchParams } from 'next/navigation'
+export const generateMetadata = async ({
+  params: { locale },
+}: MetadataProps) => {
+  const t = await getTranslator(locale, 'Login')
 
-import { useTranslations } from 'next-intl'
-
-import styles from './page.module.css'
+  return {
+    title: t('title'),
+  }
+}
 
 const LoginPage = () => {
-  const t = useTranslations('Login')
-
-  const searchParams = useSearchParams()
-  const callbackUrl = searchParams?.get('callbackUrl') ?? '/'
-
   return (
     <>
-      <div className={styles['login-container']}>
-        <div className="">
-          <span className={styles['paragraph']}>{`${t('instructions')}.`}</span>
-        </div>
-
-        <div className={styles['providers-button']}>
-          <button
-            className={`${styles['provider']} ${styles['google-provider']}`}
-            onClick={() => signIn('google', { callbackUrl })}
-          >
-            <Image
-              alt={t('Providers.Google.altImg')}
-              className={styles['image-provider']}
-              height={32}
-              src="/img/auth/google.svg"
-              width={32}
-            />
-            <span className={styles['name-provider']}>
-              {t('Providers.Google.text')}
-            </span>
-          </button>
-        </div>
-
-        <div className="">
-          <span className={styles['helper-text']}>
-            {t('HelperText.TermsOfUse.text')}{' '}
-            <Link className={styles['helper-text-link']} href="#">
-              {t('HelperText.TermsOfUse.link')}{' '}
-            </Link>
-            {t('HelperText.PrivacyPolicy.text')}{' '}
-            <Link className={styles['helper-text-link']} href="#">
-              {t('HelperText.PrivacyPolicy.link')}
-            </Link>
-          </span>
-        </div>
-      </div>
+      <Login />
     </>
   )
 }
-
 export default LoginPage
